Eliminate DiscussionService
