+++ announcements/src/org/labkey/announcements/AnnouncementModule.java
@@ -22,7 +22,6 @@ import org.labkey.announcements.config.AnnouncementEmailConfig;
 import org.labkey.announcements.model.AnnouncementDigestProvider;
 import org.labkey.announcements.model.AnnouncementManager;
 import org.labkey.announcements.model.AnnouncementType;
-import org.labkey.announcements.model.DiscussionServiceImpl;
 import org.labkey.announcements.model.InsertMessagePermission;
 import org.labkey.announcements.model.MessageBoardContributorRole;
 import org.labkey.announcements.model.SecureMessageBoardReadPermission;
@@ -30,7 +29,6 @@ import org.labkey.announcements.model.SecureMessageBoardRespondPermission;
 import org.labkey.announcements.query.AnnouncementSchema;
 import org.labkey.api.admin.FolderSerializationRegistry;
 import org.labkey.api.announcements.CommSchema;
-import org.labkey.api.announcements.DiscussionService;
 import org.labkey.api.announcements.api.AnnouncementService;
 import org.labkey.api.attachments.AttachmentService;
 import org.labkey.api.audit.AuditLogService;
@@ -98,7 +96,6 @@ public class AnnouncementModule extends DefaultModule implements SearchService.D
         AnnouncementService.setInstance(new AnnouncementServiceImpl());
 
         AnnouncementSchema.register(this);
-        DiscussionService.setInstance(new DiscussionServiceImpl());
         EmailTemplateService.get().registerTemplate(AnnouncementManager.NotificationEmailTemplate.class);
         EmailTemplateService.get().registerTemplate(AnnouncementDigestProvider.DailyDigestEmailTemplate.class);
