+++ announcements/src/org/labkey/announcements/model/SecureMessageBoardPermissions.java
@@ -18,7 +18,6 @@ package org.labkey.announcements.model;
 
 import org.jetbrains.annotations.Nullable;
 import org.labkey.api.announcements.CommSchema;
-import org.labkey.api.announcements.DiscussionService;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.security.User;
@@ -31,7 +30,7 @@ import org.labkey.api.security.permissions.ReadPermission;
  */
 public class SecureMessageBoardPermissions extends NormalMessageBoardPermissions
 {
-    public SecureMessageBoardPermissions(Container c, User user, DiscussionService.Settings settings)
+    public SecureMessageBoardPermissions(Container c, User user, Settings settings)
     {
         super(c, user, settings);
     }
